requirements.txt
librosa==0.10.2
praat-parselmouth==0.4.3
numpy==1.26.4
pandas==2.2.1
scikit-learn==1.4.1
xgboost==2.0.3
shap==0.44.1
joblib==1.3.2
sounddevice==0.4.6
scipy==1.12.0
matplotlib==3.8.3
statsmodels==0.14.1
jupyter==1.0.0
